requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
google-cloud-vision==3.4.4
openai==1.3.5
python-dotenv==1.0.0
mangum==0.17.0
Pillow==10.0.1
python-jose==3.3.0
boto3==1.28.62 